// src/app/core/models/responses/pg-details.model.ts
export interface PgDetails {
    id: string;
    name: string;
    pgType: string;
    location: string;
    city: string;
    price: number;
    image: string;
    description: string;
    features: {
      cleaning?: boolean;
      ac?: boolean;
      wifi?: boolean;
      food?: boolean;
      laundry?: boolean;
    };
    rating: number;
    contact: string;
  }
  